refactor phase 1

+++ AddicTV/Controller/FavioritesViewController.swift
@@ -17,6 +17,7 @@ class FavioritesViewController : UIViewController {
         }
         
     }
+
     
     @IBOutlet weak var tableView: UITableView!
     var context:NSManagedObjectContext {
@@ -67,30 +68,41 @@ extension FavioritesViewController:UITableViewDelegate,UITableViewDataSource {
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         print("number of rows in section \(section) is \(favorites.count)")
         
-        return favorites.count
+        return max(favorites.count, 1)
     }
-    
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return favorites.count>0 ? 80 : 160
+    }
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        print("cell for [\(indexPath.section)][\(indexPath.row)]")
+        var cell:UITableViewCell?
         
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: "cell") as? TVCell else { return UITableViewCell() }
+        if favorites.count == 0 {
+            cell = tableView.dequeueReusableCell(withIdentifier: "placeholder")
+          
+        } else {
         
-        cell.show = favorites[indexPath.row]
-        
-        
-        return cell
+            cell = tableView.dequeueReusableCell(withIdentifier: "cell") as? ShowTableViewCell
+            
+            if let tvcell = cell as? ShowTableViewCell {
+                tvcell.show = favorites[indexPath.row]
+            }
+            
+            
+        }
+        return cell!
     }
     
     func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
+        guard favorites.count>0 else { return }
         switch(editingStyle) {
         case .delete:
-            tableView.beginUpdates()
+            //tableView.beginUpdates()
             let toRemove = favorites.remove(at: indexPath.row)
             context.delete(toRemove)
             try? context.save()
-            print("deleting row at [\(indexPath.section)][\(indexPath.row)]")
-            tableView.deleteRows(at: [indexPath], with: .top)
-            tableView.endUpdates()
+            //print("deleting row at [\(indexPath.section)][\(indexPath.row)]")
+            //tableView.deleteRows(at: [indexPath], with: .top)
+            //tableView.endUpdates()
         default:
             break
         }
@@ -107,7 +119,7 @@ extension FavioritesViewController:UITableViewDelegate,UITableViewDataSource {
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         guard segue.identifier == "show favorite" else { print("wrong segue"); return }
         guard sender is TVSeries else { print("sender is not a tv series object"); return }
-        guard let svc = segue.destination as? ResultViewController else {print("wrong destination"); return}
+        guard let svc = segue.destination as? ShowViewController else {print("wrong destination"); return}
         
         let show = sender as! TVSeries
         svc.show = show
